cpu_input_boost: Introduce driver for event-based CPU boosting
This is a simple CPU input boost driver that boosts all online CPUs for
a fixed amount of time. Additionally, there is an API for other drivers
to request a boost kick (or a max-boost kick), so boosting can be done
on any custom event. This API is mainly intended for the framebuffer
driver to send a boost kick whenever there is a new frame ready to be
rendered to the display.

This driver also boosts all online CPUs to their maximum frequencies
when the display is powered on (this is the wake boost).

Since this driver requires careful tuning for optimal performance, there
are no user-exposed knobs to configure it. All necessary configuration
is done via the supplied Kconfig options.

This driver is designed for heterogeneous multi-processor systems with
two CPU clusters.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Yaroslav Furman <yaro330@gmail.com>
Signed-off-by: Panchajanya1999 <rsk52959@gmail.com>