cpufreq: stats: Drop spinlock in favor of atomic operations
CAF added cpufreq stats transition recording to the fast switch path in
commit 787ba0613fe5 ("cpufreq: record CPUFREQ stat for fast switch path"),
which has been proposed before on LKML [1] but rejected due to concerns
over the overhead of spinlocks in a scheduler hotpath.

The idea of using READ_ONCE and WRITE_ONCE instead was proposed but
dismissed due to potential synchronization issues, but atomic operations
should be safer (adding in particular) and it's highly unlikely that
we'll have multiple clusters attempting to switch their frequency at the
same time in the fast switch path. Even if that happens, we don't need
cpufreq stats to be 100% accurate so it should be a non-issue for the
most part.

[1] https://lore.kernel.org/patchwork/patch/1037238/

Test: stats look fine after running PCMark with schedutil rate-limits
      disabled
Signed-off-by: Danny Lin <danny@kdrag0n.dev>
Signed-off-by: Panchajanya1999 <panchajanya@azure-dev.live>