crypto/aes_generic: Do not align crypto_fl_tab to cacheline
Residing the 2-D array in the cache line seems to conflict with another
function with same section in shared memory.
GCC 10 caught and treats this as an fatal error and stops LTO linking.
Anyway, the difference would be negligible to run this array on a dedicated
memory region instead of a shared memory region.

Fixes (https://pastebin.com/uQ5qT7JB)

Change-Id: I18f7357ffce5683788565810ea4a2ce34d4ab2ac
Signed-off-by: Panchajanya1999 <panchajanya@azure-dev.live>